pyproject.toml
[project] 
name = "openai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.35.0",
    "openai>=1.93.2",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.2",
]
